GEOMESA-882 RemoveSchema not deleting empty z3 tables
* fixed issue where removeschema left empty z3 tables on Accumulo
* added a unit test specific to this case
* updated deleteSharedTables to check for table existance before attempting to delete anything
* modified deleteSharedTables to just drop the z3 table instead of getting a batch deleter first

+++ geomesa-accumulo/geomesa-accumulo-datastore/src/main/scala/org/locationtech/geomesa/accumulo/data/AccumuloDataStore.scala
@@ -354,9 +354,15 @@ class AccumuloDataStore(val connector: Connector,
 
     GeoMesaTable.getTables(sft).foreach { table =>
       val name = getTableName(sft.getTypeName, table)
-      val deleter = connector.createBatchDeleter(name, auths, numThreads, defaultBWConfig)
-      table.deleteFeaturesForType(sft, deleter)
-      deleter.close()
+      if (tableOps.exists(name)) {
+        if (table == Z3Table) {
+          tableOps.delete(name)
+        } else {
+          val deleter = connector.createBatchDeleter(name, auths, numThreads, defaultBWConfig)
+          table.deleteFeaturesForType(sft, deleter)
+          deleter.close()
+        }
+      }
     }
   }
